update java project

+++ pds/java/src/main/java/com/aliyun/pds/client/models/BaseCCPFileResponse.java
@@ -91,8 +91,9 @@ public class BaseCCPFileResponse extends TeaModel {
     @Validation(pattern = "[a-z0-9]{1,50}", maxLength = 50, minLength = 40)
     public String parentFileId;
 
+    // PunishFlag
     @NameInMap("punish_flag")
-    public Integer punishFlag;
+    public Long punishFlag;
 
     @NameInMap("share_id")
     public String shareId;
@@ -314,11 +315,11 @@ public class BaseCCPFileResponse extends TeaModel {
         return this.parentFileId;
     }
 
-    public BaseCCPFileResponse setPunishFlag(Integer punishFlag) {
+    public BaseCCPFileResponse setPunishFlag(Long punishFlag) {
         this.punishFlag = punishFlag;
         return this;
     }
-    public Integer getPunishFlag() {
+    public Long getPunishFlag() {
         return this.punishFlag;
     }
